+++ edbee-lib/edbee/models/chardocument/chartextdocument.h
@@ -33,9 +33,6 @@ public:
     /// This method should return the active textbuffer
     virtual TextBuffer* buffer() const;
 
-    /// this method can be used to give a 'custom' line data item to a given line
-    virtual TextLineDataManager* lineDataManager() { return textLineDataManager_; }
-
 
     /// Should return the document-scopes of this document
     virtual TextDocumentScopes* scopes() { return textScopes_; }
@@ -86,7 +83,6 @@ private:
     TextEditorConfig* config_;                               ///< The text editor configuration
     TextBuffer* textBuffer_;                                 ///< The textbuffers
 
-    TextLineDataManager* textLineDataManager_;               ///< A class for managing text line data items
     TextDocumentScopes* textScopes_;                         ///< The text document scopes
     TextLexer* textLexer_;                                   ///< The lexer used for finding the scopes